Product servisi ve product sınıfları yazıldı

+++ Product-Service/src/main/java/com/ProductService/repository/ProductRepository.java
@@ -3,5 +3,6 @@ package com.ProductService.repository;
 import com.ProductService.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
 }
